Protocol change: 16 bit annotation and layer IDs
The first 8 bits of the ID is the user's context ID. This allows each user
to make up new clash-free IDs without the server's guidance.
Predictable IDs are required for the local fork to work correctly.
Besides simplyfing the server a little, this also makes it possible to
implement features such as "paste to a new layer".

+++ src/client/core/layerstack.h
@@ -57,7 +57,7 @@ class LayerStack : public QObject {
 		void mergeLayerDown(int id);
 
 		//! Re-order the layer stack
-		void reorderLayers(const QList<uint8_t> &neworder);
+		void reorderLayers(const QList<uint16_t> &neworder);
 
 		//! Get the number of layers in the stack
 		int layers() const { return _layers.count(); }
